+++ src/ActionGameCore/Assets/FPSCameraControls/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.1.0] - 2024-03-06
+### Fixed
+- Change parameter type of input to InputActionReference.
+
 ## [1.0.1] - 2024-03-04
 ### Fixed
 - Show input action name to inspector.