+++ src/ActionGameCore/Assets/Interactions/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.5.1] - 2024-03-06
+### Fixed
+- Fix input action enabled.
+
 ## [0.5.0] - 2024-03-04
 ### Added
 - Add select event.